Авторизация, регистрация

+++ src/controllers/cards.ts
@@ -3,6 +3,7 @@ import { Error } from 'mongoose';
 import Card, { ICard } from '../models/card';
 import NotFoundError from '../helpers/errors/NotFoundError';
 import ClientError from '../helpers/errors/ClientError';
+import ForbiddenError from '../helpers/errors/ForbiddenError';
 
 export const getCards = async (req: Request, res: Response, next: NextFunction) => {
   try {
@@ -22,7 +23,7 @@ export const createCard = async (req: Request, res: Response, next: NextFunction
     res.send({ data: card });
   } catch (error) {
     if (error instanceof Error.ValidationError) {
-      throw new ClientError('Переданы некорректные данные при создании карточки');
+      next(new ClientError('Переданы некорректные данные при создании карточки'));
     } else {
       next(error);
     }
@@ -33,10 +34,16 @@ export const deleteCard = async (req: Request, res: Response, next: NextFunction
   try {
     const { cardId } = req.params;
 
+    const card: ICard | null = await Card.findById(cardId);
+
+    if (card?.owner.toString() !== req.user._id) {
+      next(new ForbiddenError('Недостаточно прав'));
+    }
+
     const result = await Card.deleteOne({ _id: cardId });
 
     if (result.deletedCount === 0) {
-      throw new NotFoundError('Карточка не найдена');
+      next(new NotFoundError('Карточка не найдена'));
     }
     res.send({ message: 'Карточка удалена' });
   } catch (error) {
@@ -55,7 +62,7 @@ export const likeCard = async (req: Request, res: Response, next: NextFunction) 
 
     const card: ICard | null = await Card
       .findByIdAndUpdate(cardId, { $addToSet: { likes: userId } }, { new: true })
-      .orFail(() => new NotFoundError('Передан несуществующий _id карточки'))
+      .orFail(() => new NotFoundError('Передан несуществующий _id карточки'));
 
     res.send({ data: card });
   } catch (error) {
@@ -74,7 +81,7 @@ export const dislikeCard = async (req: Request, res: Response, next: NextFunctio
 
     const card: ICard | null = await Card
       .findByIdAndUpdate(cardId, { $pull: { likes: userId } }, { new: true })
-      .orFail(() => new NotFoundError('Передан несуществующий _id карточки'))
+      .orFail(() => new NotFoundError('Передан несуществующий _id карточки'));
 
     res.send({ data: card });
   } catch (error) {
